Move some types around.
