+++ tsconfig-npm.json
@@ -1,5 +1,10 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["src/**/*", "src/styles/**/*", "src/indexLib.ts"],
-  "exclude": ["node_modules", "src/**/__tests__/*"]
+  "include": ["src/**/*"],
+  "exclude": [
+    "node_modules",
+    "src/stories/**/*",
+    "src/_codux/**/*",
+    "src/**/__tests__/*"
+  ]
 }
